move stitching to pred module
